refactor(api): Share filters module
